Move types used in evaluateExpression to file types.ts
